Conflicts:
	modules/activiti-bpmn-converter/src/main/java/org/activiti/bpmn/converter/BaseBpmnXMLConverter.java

+++ modules/activiti-engine/src/test/java/org/activiti/engine/test/api/event/CommentEventsTest.java
@@ -15,6 +15,7 @@ package org.activiti.engine.test.api.event;
 import org.activiti.engine.delegate.event.ActivitiEntityEvent;
 import org.activiti.engine.delegate.event.ActivitiEvent;
 import org.activiti.engine.delegate.event.ActivitiEventType;
+import org.activiti.engine.impl.history.HistoryLevel;
 import org.activiti.engine.impl.test.PluggableActivitiTestCase;
 import org.activiti.engine.runtime.ProcessInstance;
 import org.activiti.engine.task.Comment;
@@ -35,52 +36,25 @@ public class CommentEventsTest extends PluggableActivitiTestCase {
 	 */
 	@Deployment(resources = { "org/activiti/engine/test/api/runtime/oneTaskProcess.bpmn20.xml" })
 	public void testCommentEntityEvents() throws Exception {
-		ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("oneTaskProcess");
-		
-		Task task = taskService.createTaskQuery().processInstanceId(processInstance.getId()).singleResult();
-		assertNotNull(task);
-		
-		// Create link-comment
-		Comment comment = taskService.addComment(task.getId(), task.getProcessInstanceId(), "comment");
-		assertEquals(1, listener.getEventsReceived().size());
-		ActivitiEntityEvent event = (ActivitiEntityEvent) listener.getEventsReceived().get(0);
-		assertEquals(ActivitiEventType.ENTITY_CREATED, event.getType());
-		assertEquals(processInstance.getId(), event.getProcessInstanceId());
-		assertEquals(processInstance.getId(), event.getExecutionId());
-		assertEquals(processInstance.getProcessDefinitionId(), event.getProcessDefinitionId());
-		Comment commentFromEvent = (Comment) event.getEntity();
-		assertEquals(comment.getId(), commentFromEvent.getId());
-		listener.clearEventsReceived();
-		
-		// Finally, delete comment
-		taskService.deleteComment(comment.getId());
-		assertEquals(1, listener.getEventsReceived().size());
-		event = (ActivitiEntityEvent) listener.getEventsReceived().get(0);
-		assertEquals(ActivitiEventType.ENTITY_DELETED, event.getType());
-		assertEquals(processInstance.getId(), event.getProcessInstanceId());
-		assertEquals(processInstance.getId(), event.getExecutionId());
-		assertEquals(processInstance.getProcessDefinitionId(), event.getProcessDefinitionId());
-		commentFromEvent = (Comment) event.getEntity();
-		assertEquals(comment.getId(), commentFromEvent.getId());
-	}
-	
-	public void testCommentEntityEventsStandaloneTask() throws Exception {
-		Task task = null;
-		try {
-			task = taskService.newTask();
-			taskService.saveTask(task);
+		if(processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.AUDIT)) {
+			ProcessInstance processInstance = runtimeService.startProcessInstanceByKey("oneTaskProcess");
+			
+			Task task = taskService.createTaskQuery().processInstanceId(processInstance.getId()).singleResult();
 			assertNotNull(task);
 			
 			// Create link-comment
-			Comment comment = taskService.addComment(task.getId(), null, "comment");
-			assertEquals(1, listener.getEventsReceived().size());
+			Comment comment = taskService.addComment(task.getId(), task.getProcessInstanceId(), "comment");
+			assertEquals(2, listener.getEventsReceived().size());
 			ActivitiEntityEvent event = (ActivitiEntityEvent) listener.getEventsReceived().get(0);
 			assertEquals(ActivitiEventType.ENTITY_CREATED, event.getType());
-			assertNull(event.getProcessInstanceId());
-			assertNull(event.getExecutionId());
-			assertNull(event.getProcessDefinitionId());
+			assertEquals(processInstance.getId(), event.getProcessInstanceId());
+			assertEquals(processInstance.getId(), event.getExecutionId());
+			assertEquals(processInstance.getProcessDefinitionId(), event.getProcessDefinitionId());
 			Comment commentFromEvent = (Comment) event.getEntity();
 			assertEquals(comment.getId(), commentFromEvent.getId());
+			
+			event = (ActivitiEntityEvent) listener.getEventsReceived().get(1);
+			assertEquals(ActivitiEventType.ENTITY_INITIALIZED, event.getType());
 			listener.clearEventsReceived();
 			
 			// Finally, delete comment
@@ -88,16 +62,53 @@ public class CommentEventsTest extends PluggableActivitiTestCase {
 			assertEquals(1, listener.getEventsReceived().size());
 			event = (ActivitiEntityEvent) listener.getEventsReceived().get(0);
 			assertEquals(ActivitiEventType.ENTITY_DELETED, event.getType());
-			assertNull(event.getProcessInstanceId());
-			assertNull(event.getExecutionId());
-			assertNull(event.getProcessDefinitionId());
+			assertEquals(processInstance.getId(), event.getProcessInstanceId());
+			assertEquals(processInstance.getId(), event.getExecutionId());
+			assertEquals(processInstance.getProcessDefinitionId(), event.getProcessDefinitionId());
 			commentFromEvent = (Comment) event.getEntity();
 			assertEquals(comment.getId(), commentFromEvent.getId());
-			
-		} finally {
-			if(task != null && task.getId() != null) {
-				taskService.deleteTask(task.getId());
-				historyService.deleteHistoricTaskInstance(task.getId());
+		}
+	}
+	
+	public void testCommentEntityEventsStandaloneTask() throws Exception {
+		if(processEngineConfiguration.getHistoryLevel().isAtLeast(HistoryLevel.AUDIT)) {
+			Task task = null;
+			try {
+				task = taskService.newTask();
+				taskService.saveTask(task);
+				assertNotNull(task);
+				
+				// Create link-comment
+				Comment comment = taskService.addComment(task.getId(), null, "comment");
+				assertEquals(2, listener.getEventsReceived().size());
+				ActivitiEntityEvent event = (ActivitiEntityEvent) listener.getEventsReceived().get(0);
+				assertEquals(ActivitiEventType.ENTITY_CREATED, event.getType());
+				assertNull(event.getProcessInstanceId());
+				assertNull(event.getExecutionId());
+				assertNull(event.getProcessDefinitionId());
+				Comment commentFromEvent = (Comment) event.getEntity();
+				assertEquals(comment.getId(), commentFromEvent.getId());
+				
+				event = (ActivitiEntityEvent) listener.getEventsReceived().get(1);
+				assertEquals(ActivitiEventType.ENTITY_INITIALIZED, event.getType());
+				listener.clearEventsReceived();
+				
+				// Finally, delete comment
+				taskService.deleteComment(comment.getId());
+				assertEquals(1, listener.getEventsReceived().size());
+				event = (ActivitiEntityEvent) listener.getEventsReceived().get(0);
+				assertEquals(ActivitiEventType.ENTITY_DELETED, event.getType());
+				assertNull(event.getProcessInstanceId());
+				assertNull(event.getExecutionId());
+				assertNull(event.getProcessDefinitionId());
+				commentFromEvent = (Comment) event.getEntity();
+				assertEquals(comment.getId(), commentFromEvent.getId());
+				
+			} finally {
+				if(task != null && task.getId() != null) {
+					taskService.deleteTask(task.getId());
+					historyService.deleteHistoricTaskInstance(task.getId());
+				}
 			}
 		}
 	}
